requirements.txt
dash==2.0.0
plotly==5.3.1
python-binance==1.0.19
pandas==1.3.3
gunicorn==20.1.0